// src/app/models/Accounts-page/fund-receive/IFundReceiveModel.tsx
import moment from 'moment'
import {number} from 'yup'

export interface IFundReceiveModel {
  projectAmount: number
  projectCategoryName: string
  projectName: string
  paymentID: number
  projectNumber: string
  projectTypeID: number
  date: string
  remainigAmt: number
  remainingAmount: number
  cashAccountTypeID: number
  email: string
  cashAccountTypeName: string
  cashAccountID: number
  cashAccounID: number
  bankID: number
  mobileNumber: number
  receiveAmt: number
  paidAmount: number
  customerName: string
  projectStatusName: string
  isActive: boolean
  FromSubaccountID: number
  ToSubaccontID: number
  // ------ Add Fund Field------
  projectID: number
  projectInvoiceID: number
  paymentDate: string
  paymentMode: string
  fundReceiveID: number
  amount: number
  transactionModeID: number
  transactionID: string
  chequeBankName: string
  chequeBankBranch: string
  chequeDate: string
  chequeAmount: number
  chequeNumber: string
  cashAccountBankID: number
  projectInvoiceNo: string
  description: string
  filePath: string
  createBy: number
  accountBalance: number
  ipAddress: string
  cashEmployeeBalanceID: number
  cashAccontToSubID: number
  totalAmount: number
  gstAmount: number
  voucherNumber: string
  invoiceDate: string
  projectInvoiceAmount: number
  invoicePaidAmount: number
  isTcs: boolean
  tcsAmount: number
  tcsPercentage: number
  fundRecTypeID: number
  vendorTypeID: number
  payFromTypeID: number
  fromTypeID: number
}
export const FundReceiveInitValues: IFundReceiveModel = {
  projectName: '',
  FromSubaccountID: 0,
  ToSubaccontID: 0,
  projectAmount: 0,
  cashAccounID: 0,
  projectCategoryName: '',
  paymentID: 0,
  remainigAmt: 0,
  remainingAmount: 0,
  cashAccountTypeID: 0,
  cashAccountTypeName: '',
  projectNumber: '',
  projectTypeID: 0,
  fundReceiveID: 0,
  projectStatusName: '',
  date: moment(new Date()).format('YYYY-MM-DD'),
  cashAccountID: 0,
  bankID: 0,
  customerName: '',
  email: '',
  accountBalance: 0,
  mobileNumber: 0,
  receiveAmt: 0,
  paidAmount: 0,
  isActive: false,
  // ------ Add Fund Field------
  projectID: 0,
  projectInvoiceID: 0,
  paymentMode: '',
  paymentDate: moment(new Date()).format('YYYY-MM-DD'),
  amount: 0,
  transactionModeID: 0,
  transactionID: '',
  chequeBankName: '',
  chequeBankBranch: '',
  chequeDate: '',
  chequeAmount: 0,
  chequeNumber: '',
  cashAccountBankID: 0,
  projectInvoiceNo: '',
  description: '',
  filePath: '',
  createBy: 0,
  ipAddress: '',
  cashEmployeeBalanceID: 0,
  cashAccontToSubID: 0,
  totalAmount: 0,
  gstAmount: 0,
  voucherNumber: '',
  invoiceDate: moment(new Date()).format('YYYY-MM-DD'),
  projectInvoiceAmount: 0,
  invoicePaidAmount: 0,
  isTcs: false,
  tcsAmount: 0,
  tcsPercentage: 0,
  fundRecTypeID: 0,
  vendorTypeID: 0,
  payFromTypeID: 0,
  fromTypeID: 0,
}
export interface IFundRecListModel {
  projectPaymentID: number
  projectID: number
  customerID: number
  transactionModeID: number
  amount: number
  transactionID: string
  createBy: number
  createByName: string
  paymentDate: string
  projectInvoiceNo: string
  voucherNo: string
  filePath: string
  projectName: string
  customeName: string
  cashAccountName: string
  transactionMode: string
}
export interface IPaymentFromModel {
  fromID: number
  fromCode: string
  fromName: string
}
export interface IProjectDropDwnModel {
  projectID: number
  projectName: string
}